Extend JetpackTunnelGsonRequestBuilder with put and delete methods
